feat: [BBND-1498] move Dividend use cases from equity/dividends to dividend
