+++ packages/ats/sdk/__tests__/fixtures/equity/EquityFixture.ts
@@ -28,10 +28,10 @@ import GetAllScheduledBalanceAdjustmentsRequest from "@port/in/request/equity/Ge
 import { ScheduledBalanceAdjustment } from "@domain/context/equity/ScheduledBalanceAdjustment";
 import { GetScheduledBalanceAdjustmentCountQuery } from "@query/equity/balanceAdjustments/getScheduledBalanceAdjustmentCount/GetScheduledBalanceAdjustmentsCountQuery";
 import { Dividend } from "@domain/context/dividend/Dividend";
-import { GetDividendsCountQuery } from "@query/equity/dividends/getDividendsCount/GetDividendsCountQuery";
-import { GetDividendQuery } from "@query/equity/dividends/getDividend/GetDividendQuery";
-import { GetDividendForQuery } from "@query/equity/dividends/getDividendFor/GetDividendForQuery";
-import { GetDividendAmountForQuery } from "@query/equity/dividends/getDividendAmountFor/GetDividendAmountForQuery";
+import { GetDividendsCountQuery } from "@query/dividend/getDividendsCount/GetDividendsCountQuery";
+import { GetDividendQuery } from "@query/dividend/getDividend/GetDividendQuery";
+import { GetDividendForQuery } from "@query/dividend/getDividendFor/GetDividendForQuery";
+import { GetDividendAmountForQuery } from "@query/dividend/getDividendAmountFor/GetDividendAmountForQuery";
 import { DividendFor } from "@domain/context/dividend/DividendFor";
 import { GetEquityDetailsQuery } from "@query/equity/get/getEquityDetails/GetEquityDetailsQuery";
 import { EquityDetails } from "@domain/context/equity/EquityDetails";
@@ -47,16 +47,16 @@ import { SetScheduledBalanceAdjustmentCommand } from "@command/equity/balanceAdj
 import CancelScheduledBalanceAdjustmentRequest from "@port/in/request/equity/CancelScheduledBalanceAdjustmentRequest";
 import { CancelScheduledBalanceAdjustmentCommand } from "@command/equity/balanceAdjustments/cancelScheduledBalanceAdjustment/CancelScheduledBalanceAdjustmentCommand";
 import { CreateEquityCommand } from "@command/equity/create/CreateEquityCommand";
-import { SetDividendCommand } from "@command/equity/dividends/set/SetDividendCommand";
-import { CancelDividendCommand } from "@command/equity/dividends/cancel/CancelDividendCommand";
+import { SetDividendCommand } from "@command/dividend/set/SetDividendCommand";
+import { CancelDividendCommand } from "@command/dividend/cancel/CancelDividendCommand";
 import CancelDividendRequest from "@port/in/request/dividend/CancelDividendRequest";
 import { SetVotingRightsCommand } from "@command/equity/votingRights/set/SetVotingRightsCommand";
 import { CancelVotingCommand } from "@command/equity/votingRights/cancel/CancelVotingCommand";
 import CancelVotingRequest from "@port/in/request/equity/CancelVotingRequest";
 import ContractId from "@domain/context/contract/ContractId";
 import { SecurityPropsFixture } from "../shared/SecurityFixture";
-import { GetDividendHoldersQuery } from "@query/equity/dividends/getDividendHolders/GetDividendHoldersQuery";
-import { GetTotalDividendHoldersQuery } from "@query/equity/dividends/getTotalDividendHolders/GetTotalDividendHoldersQuery";
+import { GetDividendHoldersQuery } from "@query/dividend/getDividendHolders/GetDividendHoldersQuery";
+import { GetTotalDividendHoldersQuery } from "@query/dividend/getTotalDividendHolders/GetTotalDividendHoldersQuery";
 import { GetVotingHoldersQuery } from "@query/equity/votingRights/getVotingHolders/GetVotingHoldersQuery";
 import { GetTotalVotingHoldersQuery } from "@query/equity/votingRights/getTotalVotingHolders/GetTotalVotingHoldersQuery";
 import GetDividendHoldersRequest from "@port/in/request/dividend/GetDividendHoldersRequest";
