+++ packages/ats/sdk/src/app/usecase/command/dividend/cancel/error/CancelDividendCommandError.ts
