+++ packages/ats/sdk/src/app/usecase/command/dividend/set/SetDividendCommand.ts
