+++ packages/ats/sdk/src/app/usecase/command/dividend/set/error/SetDividendCommandError.ts
