+++ packages/ats/sdk/src/app/usecase/query/dividend/getDividend/GetDividendQuery.ts
