+++ packages/ats/sdk/src/app/usecase/query/dividend/getDividend/GetDividendQueryHandler.ts
