+++ packages/ats/sdk/src/app/usecase/query/dividend/getDividend/error/GetDividendQueryError.ts
