+++ packages/ats/sdk/src/app/usecase/query/dividend/getDividendAmountFor/GetDividendAmountForQuery.ts
