+++ packages/ats/sdk/src/app/usecase/query/dividend/getDividendAmountFor/GetDividendAmountForQueryHandler.ts
