+++ packages/ats/sdk/src/app/usecase/query/dividend/getDividendAmountFor/error/GetDividendAmountForQueryError.ts
