+++ packages/ats/sdk/src/app/usecase/query/dividend/getDividendFor/GetDividendForQueryHandler.ts
