+++ packages/ats/sdk/src/app/usecase/query/dividend/getDividendFor/error/GetDividendForQueryError.ts
