+++ packages/ats/sdk/src/app/usecase/query/dividend/getDividendHolders/error/GetDividendHoldersQueryError.ts
