+++ packages/ats/sdk/src/app/usecase/query/dividend/getDividendsCount/GetDividendsCountQuery.ts
