+++ packages/ats/sdk/src/app/usecase/query/dividend/getDividendsCount/GetDividendsCountQueryHandler.unit.test.ts
