+++ packages/ats/sdk/src/app/usecase/query/dividend/getDividendsCount/error/GetDividendsCountQueryError.ts
