+++ packages/ats/sdk/src/app/usecase/query/dividend/getTotalDividendHolders/GetTotalDividendHoldersQuery.ts
