+++ packages/ats/sdk/src/app/usecase/query/dividend/getTotalDividendHolders/GetTotalDividendHoldersQueryHandler.ts
