+++ packages/ats/sdk/src/app/usecase/query/dividend/getTotalDividendHolders/error/GetTotalDividendHoldersQueryError.ts
