+++ packages/ats/sdk/src/core/injectable/equity/InjectableEquity.ts
@@ -2,20 +2,20 @@
 
 import { TOKENS } from "../Tokens";
 import { CreateEquityCommandHandler } from "@command/equity/create/CreateEquityCommandHandler";
-import { SetDividendCommandHandler } from "@command/equity/dividends/set/SetDividendCommandHandler";
-import { CancelDividendCommandHandler } from "@command/equity/dividends/cancel/CancelDividendCommandHandler";
+import { SetDividendCommandHandler } from "@command/dividend/set/SetDividendCommandHandler";
+import { CancelDividendCommandHandler } from "@command/dividend/cancel/CancelDividendCommandHandler";
 import { SetVotingRightsCommandHandler } from "@command/equity/votingRights/set/SetVotingRightsCommandHandler";
 import { CancelVotingCommandHandler } from "@command/equity/votingRights/cancel/CancelVotingCommandHandler";
-import { GetDividendForQueryHandler } from "@query/equity/dividends/getDividendFor/GetDividendForQueryHandler";
-import { GetDividendAmountForQueryHandler } from "@query/equity/dividends/getDividendAmountFor/GetDividendAmountForQueryHandler";
-import { GetDividendQueryHandler } from "@query/equity/dividends/getDividend/GetDividendQueryHandler";
-import { GetDividendsCountQueryHandler } from "@query/equity/dividends/getDividendsCount/GetDividendsCountQueryHandler";
+import { GetDividendForQueryHandler } from "@query/dividend/getDividendFor/GetDividendForQueryHandler";
+import { GetDividendAmountForQueryHandler } from "@query/dividend/getDividendAmountFor/GetDividendAmountForQueryHandler";
+import { GetDividendQueryHandler } from "@query/dividend/getDividend/GetDividendQueryHandler";
+import { GetDividendsCountQueryHandler } from "@query/dividend/getDividendsCount/GetDividendsCountQueryHandler";
 import { GetVotingForQueryHandler } from "@query/equity/votingRights/getVotingFor/GetVotingForQueryHandler";
 import { GetVotingQueryHandler } from "@query/equity/votingRights/getVoting/GetVotingQueryHandler";
 import { GetVotingCountQueryHandler } from "@query/equity/votingRights/getVotingCount/GetVotingCountQueryHandler";
 import { GetEquityDetailsQueryHandler } from "@query/equity/get/getEquityDetails/GetEquityDetailsQueryHandler";
-import { GetDividendHoldersQueryHandler } from "@query/equity/dividends/getDividendHolders/GetDividendHoldersQueryHandler";
-import { GetTotalDividendHoldersQueryHandler } from "@query/equity/dividends/getTotalDividendHolders/GetTotalDividendHoldersQueryHandler";
+import { GetDividendHoldersQueryHandler } from "@query/dividend/getDividendHolders/GetDividendHoldersQueryHandler";
+import { GetTotalDividendHoldersQueryHandler } from "@query/dividend/getTotalDividendHolders/GetTotalDividendHoldersQueryHandler";
 import { GetVotingHoldersQueryHandler } from "@query/equity/votingRights/getVotingHolders/GetVotingHoldersQueryHandler";
 import { GetTotalVotingHoldersQueryHandler } from "@query/equity/votingRights/getTotalVotingHolders/GetTotalVotingHoldersQueryHandler";
 import { CancelScheduledBalanceAdjustmentCommandHandler } from "@command/equity/balanceAdjustments/cancelScheduledBalanceAdjustment/CancelScheduledBalanceAdjustmentCommandHandler";
