+++ packages/ats/sdk/src/port/in/equity/Equity.unit.test.ts
@@ -76,17 +76,17 @@ import { CancelVotingCommand } from "@command/equity/votingRights/cancel/CancelV
 import { GetVotingForQuery } from "@query/equity/votingRights/getVotingFor/GetVotingForQuery";
 import { GetVotingQuery } from "@query/equity/votingRights/getVoting/GetVotingQuery";
 import { GetVotingCountQuery } from "@query/equity/votingRights/getVotingCount/GetVotingCountQuery";
-import { SetDividendCommand } from "@command/equity/dividends/set/SetDividendCommand";
-import { CancelDividendCommand } from "@command/equity/dividends/cancel/CancelDividendCommand";
-import { GetDividendForQuery } from "@query/equity/dividends/getDividendFor/GetDividendForQuery";
-import { GetDividendQuery } from "@query/equity/dividends/getDividend/GetDividendQuery";
-import { GetDividendsCountQuery } from "@query/equity/dividends/getDividendsCount/GetDividendsCountQuery";
+import { SetDividendCommand } from "@command/dividend/set/SetDividendCommand";
+import { CancelDividendCommand } from "@command/dividend/cancel/CancelDividendCommand";
+import { GetDividendForQuery } from "@query/dividend/getDividendFor/GetDividendForQuery";
+import { GetDividendQuery } from "@query/dividend/getDividend/GetDividendQuery";
+import { GetDividendsCountQuery } from "@query/dividend/getDividendsCount/GetDividendsCountQuery";
 import { SetScheduledBalanceAdjustmentCommand } from "@command/equity/balanceAdjustments/setScheduledBalanceAdjustment/SetScheduledBalanceAdjustmentCommand";
 import { GetScheduledBalanceAdjustmentQuery } from "@query/equity/balanceAdjustments/getScheduledBalanceAdjustment/GetScheduledBalanceAdjustmentQuery";
 import { GetScheduledBalanceAdjustmentCountQuery } from "@query/equity/balanceAdjustments/getScheduledBalanceAdjustmentCount/GetScheduledBalanceAdjustmentsCountQuery";
 import { CancelScheduledBalanceAdjustmentCommand } from "@command/equity/balanceAdjustments/cancelScheduledBalanceAdjustment/CancelScheduledBalanceAdjustmentCommand";
-import { GetDividendHoldersQuery } from "@query/equity/dividends/getDividendHolders/GetDividendHoldersQuery";
-import { GetTotalDividendHoldersQuery } from "@query/equity/dividends/getTotalDividendHolders/GetTotalDividendHoldersQuery";
+import { GetDividendHoldersQuery } from "@query/dividend/getDividendHolders/GetDividendHoldersQuery";
+import { GetTotalDividendHoldersQuery } from "@query/dividend/getTotalDividendHolders/GetTotalDividendHoldersQuery";
 import { GetVotingHoldersQuery } from "@query/equity/votingRights/getVotingHolders/GetVotingHoldersQuery";
 import { GetTotalVotingHoldersQuery } from "@query/equity/votingRights/getTotalVotingHolders/GetTotalVotingHoldersQuery";
 import { CreateTrexSuiteEquityCommand } from "@command/equity/createTrexSuite/CreateTrexSuiteEquityCommand";
